refactor(media-editor): split camera.ts, extract hooks, trim cropper.tsx
Address PR review feedback on file size and organization:

- Split core/camera.ts (934 lines) into three cohesive files:
  - camera.ts — primitives: createCamera, createExportCamera,
    worldToScreen, screenToWorld, getVisibleBounds, getImageFit,
    getRotatedBBox.
  - containment.ts — restriction source of truth: restrictPanZoom,
    restrictCropRect, getCropBounds, plus the private
    getVisualDimensions / getMinZoomForCover helpers.
  - source-region.ts — pixel-region queries: getSourceRegion,
    getSourceRegionPercent and their types.

- Extract the ARIA-live announcement logic in cropper.tsx into a
  dedicated useAriaAnnouncer hook. Reduces the component by ~30
  lines and makes the debounce/dedupe surface reusable.

- DRY the two aspect-ratio auto-resize effects in cropper.tsx by
  factoring their shared math into a computeInscribedRect helper.
  The effects stay in place because they have different triggers
  (fixed vs. freeform modes).

- Simplify InteractionController.handleTouchStart by moving the
  inline onTouchMove closure (~170 lines) into a bound
  handleTouchMove arrow-property, and extracting double-tap
  detection into a tryDoubleTap method.

- Sync stale references in docs / comments after the moves
  (architecture.md diagram, createExportCamera docstring, etc.).

Net line counts: camera.ts 934→357, cropper.tsx 559→497, plus
cohesive new files. No behavior change — all 12,458 unit tests
(including the preview↔export parity sweep) still pass.
